refactor undo_server as simplified observer_server
